// linux/northstar/ui/src/gen/northstar_pb.ts
export class Service extends Message<Service> {
  /**
   * @generated from field: uint32 id = 1;
   */
  id = 0;

  /**
   * @generated from field: uint32 host_id = 2;
   */
  hostId = 0;

  /**
   * @generated from field: string name = 3;
   */
  name = "";

  /**
   * @generated from field: string technology = 4;
   */
  technology = "";

  /**
   * @generated from field: bool scored = 5;
   */
  scored = false;

  /**
   * @generated from field: bool disabled = 6;
   */
  disabled = false;

  /**
   * @generated from field: bool backed_up = 7;
   */
  backedUp = false;

  /**
   * @generated from field: bool hardened = 8;
   */
  hardened = false;

  /**
   * @generated from field: bool ldap_authentication = 9;
   */
  ldapAuthentication = false;

  /**
   * @generated from field: repeated northstar.ServicePort service_ports = 10;
   */
  servicePorts: ServicePort[] = [];

  /**
   * @generated from field: repeated northstar.Website websites = 11;
   */
  websites: Website[] = [];

  /**
   * @generated from field: optional int32 password_index = 12;
   */
  passwordIndex?: number;

  /**
   * @generated from field: repeated northstar.ServiceDependency dependencies = 13;
   */
  dependencies: ServiceDependency[] = [];

  constructor(data?: PartialMessage<Service>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = "northstar.Service";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 1, name: "id", kind: "scalar", T: 13 /* ScalarType.UINT32 */ },
    { no: 2, name: "host_id", kind: "scalar", T: 13 /* ScalarType.UINT32 */ },
    { no: 3, name: "name", kind: "scalar", T: 9 /* ScalarType.STRING */ },
    { no: 4, name: "technology", kind: "scalar", T: 9 /* ScalarType.STRING */ },
    { no: 5, name: "scored", kind: "scalar", T: 8 /* ScalarType.BOOL */ },
    { no: 6, name: "disabled", kind: "scalar", T: 8 /* ScalarType.BOOL */ },
    { no: 7, name: "backed_up", kind: "scalar", T: 8 /* ScalarType.BOOL */ },
    { no: 8, name: "hardened", kind: "scalar", T: 8 /* ScalarType.BOOL */ },
    { no: 9, name: "ldap_authentication", kind: "scalar", T: 8 /* ScalarType.BOOL */ },
    { no: 10, name: "service_ports", kind: "message", T: ServicePort, repeated: true },
    { no: 11, name: "websites", kind: "message", T: Website, repeated: true },
    { no: 12, name: "password_index", kind: "scalar", T: 5 /* ScalarType.INT32 */, opt: true },
    { no: 13, name: "dependencies", kind: "message", T: ServiceDependency, repeated: true },
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): Service {
    return new Service().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): Service {
    return new Service().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): Service {
    return new Service().fromJsonString(jsonString, options);
  }

  static equals(a: Service | PlainMessage<Service> | undefined, b: Service | PlainMessage<Service> | undefined): boolean {
    return proto3.util.equals(Service, a, b);
  }
}

/**
 * @generated from message northstar.ServiceDependency
 */
export class ServiceDependency extends Message<ServiceDependency> {
  /**
   * @generated from field: uint32 service_id = 1;
   */
  serviceId = 0;

  /**
   * @generated from field: uint32 depends_on_id = 2;
   */
  dependsOnId = 0;

  /**
   * @generated from field: string depends_on_name = 3;
   */
  dependsOnName = "";

  /**
   * @generated from field: string depends_on_host = 4;
   */
  dependsOnHost = "";

  constructor(data?: PartialMessage<ServiceDependency>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = "northstar.ServiceDependency";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 1, name: "service_id", kind: "scalar", T: 13 /* ScalarType.UINT32 */ },
    { no: 2, name: "depends_on_id", kind: "scalar", T: 13 /* ScalarType.UINT32 */ },
    { no: 3, name: "depends_on_name", kind: "scalar", T: 9 /* ScalarType.STRING */ },
    { no: 4, name: "depends_on_host", kind: "scalar", T: 9 /* ScalarType.STRING */ },
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): ServiceDependency {
    return new ServiceDependency().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): ServiceDependency {
    return new ServiceDependency().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): ServiceDependency {
    return new ServiceDependency().fromJsonString(jsonString, options);
  }

  static equals(a: ServiceDependency | PlainMessage<ServiceDependency> | undefined, b: ServiceDependency | PlainMessage<ServiceDependency> | undefined): boolean {
    return proto3.util.equals(ServiceDependency, a, b);
  }
}
// ...
export class ServicePort extends Message<ServicePort> {
  /**
   * @generated from field: uint32 service_id = 1;
   */
  serviceId = 0;

  /**
   * @generated from field: uint32 port = 2;
   */
  port = 0;

  constructor(data?: PartialMessage<ServicePort>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = "northstar.ServicePort";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 1, name: "service_id", kind: "scalar", T: 13 /* ScalarType.UINT32 */ },
    { no: 2, name: "port", kind: "scalar", T: 13 /* ScalarType.UINT32 */ },
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): ServicePort {
    return new ServicePort().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): ServicePort {
    return new ServicePort().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): ServicePort {
    return new ServicePort().fromJsonString(jsonString, options);
  }

  static equals(a: ServicePort | PlainMessage<ServicePort> | undefined, b: ServicePort | PlainMessage<ServicePort> | undefined): boolean {
    return proto3.util.equals(ServicePort, a, b);
  }
}

/**
 * @generated from message northstar.Website
 */
export class Website extends Message<Website> {
  /**
   * @generated from field: uint32 id = 1;
   */
  id = 0;

  /**
   * @generated from field: uint32 service_id = 2;
   */
  serviceId = 0;

  /**
   * @generated from field: string name = 3;
   */
  name = "";

  /**
   * @generated from field: string url = 4;
   */
  url = "";

  /**
   * @generated from field: string username = 5;
   */
  username = "";

  /**
   * @generated from field: optional int32 password_index = 6;
   */
  passwordIndex?: number;

  /**
   * @generated from field: string old_password = 7;
   */
  oldPassword = "";

  /**
   * @generated from field: bool enumerated = 8;
   */
  enumerated = false;

  constructor(data?: PartialMessage<Website>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = "northstar.Website";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 1, name: "id", kind: "scalar", T: 13 /* ScalarType.UINT32 */ },
    { no: 2, name: "service_id", kind: "scalar", T: 13 /* ScalarType.UINT32 */ },
    { no: 3, name: "name", kind: "scalar", T: 9 /* ScalarType.STRING */ },
    { no: 4, name: "url", kind: "scalar", T: 9 /* ScalarType.STRING */ },
    { no: 5, name: "username", kind: "scalar", T: 9 /* ScalarType.STRING */ },
    { no: 6, name: "password_index", kind: "scalar", T: 5 /* ScalarType.INT32 */, opt: true },
    { no: 7, name: "old_password", kind: "scalar", T: 9 /* ScalarType.STRING */ },
    { no: 8, name: "enumerated", kind: "scalar", T: 8 /* ScalarType.BOOL */ },
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): Website {
    return new Website().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): Website {
    return new Website().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): Website {
    return new Website().fromJsonString(jsonString, options);
  }

  static equals(a: Website | PlainMessage<Website> | undefined, b: Website | PlainMessage<Website> | undefined): boolean {
    return proto3.util.equals(Website, a, b);
  }
}